Added debugging options to find out Ocra issues

+++ build.rb
@@ -1,5 +1,8 @@
 require 'zip'
 
+DEBUG_MODE = true
+DEBUG_MODE ? DEBUG_CMD = "--debug-extract" : DEBUG_CMD = ""
+
 OUT_EXE = "build/yamlomiser.exe"
 OUT_ZIP = "build/yamlomiser.zip"
 
@@ -9,7 +12,7 @@ begin
   File.delete(OUT_ZIP) if File.exist?(OUT_ZIP)
 
   # builds .exe file & cleans after itself
-  system("ocra --console --no-dep-run --dll ruby_builtin_dlls/libgmp-10.dll --gem-full=yaml --verbose --output #{OUT_EXE} --icon build/rand.png randomiser.rb")
+  system("ocra --console --no-dep-run #{DEBUG_CMD} --dll ruby_builtin_dlls/libgmp-10.dll --gem-full=yaml --verbose --output #{OUT_EXE} --icon build/rand.png randomiser.rb")
   File.delete("tmpin") if File.exist?("tmpin")
 
   # packs all files into build folder
@@ -17,7 +20,9 @@ begin
     zipfile.add("values.yaml",    File.join("values.yaml"))
     zipfile.add("yamlomiser.exe", File.join(OUT_EXE))
   end
-  File.delete(OUT_EXE) if File.exist?(OUT_EXE)
+  if !DEBUG_MODE
+    File.delete(OUT_EXE) if File.exist?(OUT_EXE)
+  end
 
 rescue
   puts "Build failed."
